Refactor infrastructure details
